Shopify Build Setup
I've manually altered this in master, because it was initially committed
as multiple commits, and I would like to squash it into a single commit.

The reason a single commit is necessary is that in order to enable
branch builds on specific tags, we will need to cherry-pick this commit
ontop of each tag. Once this is a single commit it should be a scripted
process.